requirements.txt
mouseinfo==0.1.3
pillow==10.3.0
pyautogui==0.9.54
pygetwindow==0.0.9
pymsgbox==1.0.9
pyperclip==1.8.2
pyrect==0.2.0
pyscreeze==0.1.30
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
six==1.16.0
